pyproject.toml
[tool.poetry]
name = "deep_next"
version = "0.3.4"
description = "DeepNext monorepo."
authors = [
    "Bartłomiej Roszak <bartlomiej.roszak@stxnext.pl>",
    "Patryk Laskowski <patryk.laskowski@stxnext.pl>",
    "Patryk Jedlikowski <patryk.jedlikowski@stxnext.pl>",
    "Filip Szmid <filip.szmid@stxnext.pl>",
    "Daniel Iwanicki <daniel.iwanicki@stxnext.pl>",
]
package-mode = false

[tool.poetry.dependencies]
python = ">=3.11.0,<3.12.0"
click = "^8.1.7"
loguru = "^0.7.2"
python-dotenv = "^1.0.1"
pydantic = "^2.10.3"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.4"
pytest-rerunfailures = "^15.0"
pre-commit = "^4.0.1"

[tool.poetry.group.app.dependencies]
deep_next_app = { path = "apps/app", develop = true  }
deep_next_core = { path = "libs/core", develop = true  }
deep_next_connectors = { path = "libs/connectors", develop = true  }
deep_next_common = { path = "libs/common", develop = true  }

[tool.poetry.group.core.dependencies]
deep_next_core = { path = "libs/core", develop = true  }
deep_next_common = { path = "libs/common", develop = true  }
langchain = "^0.3.21"
levenshtein = "^0.26.1"
langgraph = "^0.3.20"
langchain-aws = "^0.2.9"
langchain-openai = "^0.3.10"
langchain-community = "^0.3.20"
rich = "^13.9.4"
pylint = "^3.0.0"
dill = "^0.3.9"  # pylint
sortedcontainers = "^2.4.0"
scipy = "^1.15.2"

[tool.poetry.group.connectors.dependencies]
deep_next_connectors = { path = "libs/connectors", develop = true  }
python-gitlab = "^5.1.0"
boto3 = "^1.35.80"
slack-sdk = "^3.34.0"

[tool.poetry.group.common.dependencies]
deep_next_common = { path = "libs/common", develop = true  }

[tool.pytest.ini_options]
norecursedirs = [
    "apps/*/tests/**/_resources",
    "libs/*/tests/**/_resources",
    "**/__*",
]
filterwarnings = [
    "ignore"
]
